Formatting fixes and move router.go to routers/
