+++ integrations/sqlite.ini
@@ -52,5 +52,7 @@ LEVEL     = Info
 ROOT_PATH = log
 
 [security]
-INSTALL_LOCK = true
-SECRET_KEY   = 9pCviYTWSb
+INSTALL_LOCK   = true
+SECRET_KEY     = 9pCviYTWSb
+INTERNAL_TOKEN = eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYmYiOjE0OTI3OTU5ODN9.OQkH5UmzID2XBdwQ9TAI6Jj2t1X-wElVTjbE7aoN4I8
+
